requirement.txt
torch==2.6.0
torchvision==0.21.0
transformers==4.57.3
accelerate==1.4.0
diffusers==0.36.0
tqdm==4.67.1
pillow==10.4.0
peft==0.17.0